centralize service protocols
