+++ ripper/ripperlib/defs.py
@@ -1,0 +1,24 @@
+from beartype.typing import Any, Dict, Protocol
+from typing_extensions import runtime_checkable
+
+
+# Protocol classes for Google API services
+@runtime_checkable
+class SheetsService(Protocol):
+    def spreadsheets(self) -> Any: ...
+    def values(self) -> Any: ...
+    def get(self, spreadsheetId: str) -> Any: ...
+    def batchUpdate(self, spreadsheetId: str, body: Dict[str, Any]) -> Any: ...
+
+
+@runtime_checkable
+class DriveService(Protocol):
+    def files(self) -> Any: ...
+    def get(self, fileId: str) -> Any: ...
+    def list(self, **kwargs: Any) -> Any: ...
+
+
+@runtime_checkable
+class UserInfoService(Protocol):
+    def userinfo(self) -> Any: ...
+    def get(self) -> Any: ...
